refactor(RandomUtils): 클래스명 변경 및 상수화
